+++ src/main/java/game_engine/model/entities/Enemy.java
@@ -1,4 +1,4 @@
-package game_engine.model;
+package game_engine.model.entities;
 
 import java.awt.*;
