+++ src/main/java/game_engine/model/events/EventTypes.java
@@ -1,0 +1,16 @@
+package game_engine.model.events;
+
+
+/**
+ * Enum contains all types of events.
+ *
+ * @author  Christian-2003
+ */
+public enum EventTypes {
+
+    /**
+     * Type for collision events.
+     */
+    COLLISION
+
+}
