+++ src/main/java/game_engine/model/map/objects/MapObject.java
@@ -1,4 +1,8 @@
-package game_engine.model;
+package game_engine.model.map.objects;
+
+import game_engine.model.GameObject;
+import game_engine.model.map.GameChunk;
+import game_engine.model.map.GameMap;
 
 import java.awt.*;
