+++ src/test/java/game_engine/model/GameChunkTest.java
@@ -1,9 +1,10 @@
 package game_engine.model;
 
+import game_engine.model.map.GameChunk;
+import game_engine.model.map.objects.MapObject;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import java.awt.*;
 
 
 /**
